refactor: schema region consolidation + playback cap + deep-agent fold
